974 Kaprekar Numbers, minor fixes

+++ algorithms/onlinejudge/maths/Kaprekar.cpp
@@ -37,9 +37,9 @@ namespace algorithms::onlinejudge::maths::Kaprekar
          for(int d = 1; d <= (int)sq_s.size(); ++d) {
           auto l_s = sq_s.substr(0, sq_s.size() - d);
           auto r_s = sq_s.substr(sq_s.size() - d);
-          ll l = 0, r = 0;
-          if(!l_s.empty()) l = std::stoll(l_s);
-          if(!r_s.empty()) r = std::stoll(r_s);
+          int l = 0, r = 0;
+          if(!l_s.empty()) l = std::stoi(l_s);
+          if(!r_s.empty()) r = std::stoi(r_s);
           if(l > 0 && r > 0 && l + r == i) {
             kaprekar[i] = true;
             break;
@@ -55,10 +55,10 @@ namespace algorithms::onlinejudge::maths::Kaprekar
          scanf("%d %d", &l, &r);
          bool found = false;
          for(int i = l; i <= r; ++i) {
-            if(kaprekar[i]) {
-              printf("%d\n", i);
-              if(!found) found = true;
-            }
+           if(kaprekar[i]) {
+             printf("%d\n", i);
+             if(!found) found = true;
+           }
          }
         if(!found) printf("no kaprekar numbers\n");
         if(t_cases) printf("\n");
